+++ StoreManagement/Class/Client_PackedItemClassified2.cs
@@ -12,7 +12,7 @@ namespace StoreManagement.Class
         {
         }
 
-        public Client_PackedItemClassified2(Accessory accessory, Pack pack, PackedItem packedItem, int sample, int defectLimit, int sumIdentifiedQuantity, int sumCountedQuantity, int sumCheckedQuantity,int sumUnqualifiedQuantity)
+        public Client_PackedItemClassified2(Accessory accessory, Pack pack, PackedItem packedItem, double sample, double defectLimit, double sumIdentifiedQuantity, double sumCountedQuantity, double sumCheckedQuantity, double sumUnqualifiedQuantity)
         {
             PackedItemPK = packedItem.PackedItemPK;
             AccessoryID = accessory.AccessoryID;
@@ -46,18 +46,18 @@ namespace StoreManagement.Class
 
         public string PackID { get; set; }
 
-        public int PackedQuantity { get; set; }
+        public double PackedQuantity { get; set; }
 
-        public int Sample { get; set; }
+        public double Sample { get; set; }
 
-        public int DefectLimit { get; set; }
+        public double DefectLimit { get; set; }
 
-        public int SumIdentifiedQuantity { get; set; }
+        public double SumIdentifiedQuantity { get; set; }
 
-        public int SumCountedQuantity { get; set; }
+        public double SumCountedQuantity { get; set; }
 
-        public int SumCheckedQuantity { get; set; }
+        public double SumCheckedQuantity { get; set; }
 
-        public int SumUnqualifiedQuantity { get; set; }
+        public double SumUnqualifiedQuantity { get; set; }
     }
 }
